Add switch for favorite color with alerts
Problem
Write a switch to parse input, then check
if it's my favorite color, then alert
the user if they guessed my favorite color

Solution
Wrote a simple switch that takes a number of colors.
It then checks if the input matches the goal,
and alerts the user on the page.

Note
Used Mozilla docs on Switch statements