Reorder while loop for guessing game
My while loop would run indefinitely, and certain blocks of code
would never run.

With the help of a TA, we worked to reorder and debug a while loop
built for checking guesses in a number game. We had to change the
conditionals on line 3 and 6. By setting guesses === 1 it's ensured
to run at the last guess. Added a break to line 11, so that if the
user guesses correctly the loop stops.

Check if guesses > 0 not guesses !== 0. Guesses will go past 0.
Order conditionals in least to most specific. That way all of
the conditions get checked.
While loops can use a break statement.